+++ src/sensors/pressure.rs
@@ -41,7 +41,7 @@ impl Storable for SeeedWaterPressureSensor {
 }
 
 impl PressureProbe for SeeedWaterPressureSensor {
-    fn convert_voltage_to_pressure(&self, voltage: f32) -> Result<f32, String> {
-        Ok((voltage / self.vcc - 0.1) / 0.75)
+    fn convert_voltage_to_pressure(&self, voltage: f64) -> Result<f32, String> {
+        Ok(((voltage / self.vcc as f64 - 0.1) / 0.75) as f32)
     }
 }
